Also move `rangeStep` to `src/compile/scale/range`
